extract `places_conflict` into its own module
